refactor(runtime): extract history config module from agent_loop
The message-history trim config — `DEFAULT_MAX_HISTORY_MESSAGES`
(pub, 60), `MIN_HISTORY_MESSAGES` (pub(super), 4), `resolve_max_history`
(pub(super)) and `clamp_max_history` (private) — moves into
`agent_loop/history.rs`.

`resolve_max_history` is used by the main `run_agent_loop` and
`run_agent_loop_streaming` entry points to derive the per-turn trim
cap; `DEFAULT_MAX_HISTORY_MESSAGES` is re-exported via
`pub use self::history::DEFAULT_MAX_HISTORY_MESSAGES` so docs that
quote `librefang_runtime::agent_loop::DEFAULT_MAX_HISTORY_MESSAGES`
(message-history-trimming architecture page, configuration core page)
keep resolving. `MIN_HISTORY_MESSAGES` is referenced from the in-file
`mod tests` clamp-tests; pulled in alongside the existing
`super::model::*` / `super::web_augment::*` test-only imports.

`clamp_max_history` stays private — `resolve_max_history` is its only
caller. The 4-tier resolution order
(`manifest > opts > DEFAULT > clamped-to-floor`) is preserved exactly.

Behavior-preserving. mod.rs: 13,096 -> 13,027 LOC (-69).
cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
